// src/lib/supabase/types.d.ts
export type Database = {
// ...
  public: {
    Tables: {
      favorite_shogi_moves: {
        Row: {
          created_at: string
          id: number
          move: string
          sfenx: string
          user_id: string | null
        }
        Insert: {
          created_at?: string
          id?: number
          move?: string
          sfenx?: string
          user_id?: string | null
        }
        Update: {
          created_at?: string
          id?: number
          move?: string
          sfenx?: string
          user_id?: string | null
        }
        Relationships: []
      }
      game_records: {
        Row: {
          black_player: string | null
          created_at: string
          end_time: string | null
          event: string | null
          game_hash: string
          handicap: string | null
          id: number
          move_count: number
          recorded_at: string
          result: string | null
          start_time: string | null
          user_id: string | null
          white_player: string | null
        }
        Insert: {
          black_player?: string | null
          created_at?: string
          end_time?: string | null
          event?: string | null
          game_hash: string
          handicap?: string | null
          id?: number
          move_count: number
          recorded_at?: string
          result?: string | null
          start_time?: string | null
          user_id?: string | null
          white_player?: string | null
        }
        Update: {
          black_player?: string | null
          created_at?: string
          end_time?: string | null
          event?: string | null
          game_hash?: string
          handicap?: string | null
          id?: number
          move_count?: number
          recorded_at?: string
          result?: string | null
          start_time?: string | null
          user_id?: string | null
          white_player?: string | null
        }
        Relationships: []
      }
      shogi_moves_statistics: {
        Row: {
          created_at: string
          id: number
          lose: boolean
          move: string
          sfenx: string
          timeout: boolean
          user_id: string | null
          win: boolean
        }
        Insert: {
          created_at?: string
          id?: number
          lose: boolean
          move?: string
          sfenx?: string
          timeout: boolean
          user_id?: string | null
          win: boolean
        }
        Update: {
          created_at?: string
          id?: number
          lose?: boolean
          move?: string
          sfenx?: string
          timeout?: boolean
          user_id?: string | null
          win?: boolean
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never